Use only Ecommerce stored payment methods
In the paymentMethods response since the checkout version 51 we get back
the storedPaymentMethods object which includes both Ecommerce and
ContAuth stored payment methods. Before we had only the
oneClickPaymentMethods which only included the Ecommerce ones.
If a storedPaymentMethod doesn't have Ecommerce in it's
supportedShopperInteractions array, then don't use it as a oneClick
method.