Treat a `Refused` error code for 3DS process
As of now, our PrestaShop plugin would show a blank page with a JSON with the error message. This is because the back-end would send an empty response with a 200 status code, tricking the front-end into treating this as a successful payment.

Changing the action of this response to `error` will inform the front-end how to properly handle this issue.