Fix open invoice line item for shipping costs
The current calculation sends information as a floating point number, which is incorrect and prone to issues when dealing with non-exact costs (€8.05, for example).

To fix this, we are formatting the number we get as shipping costs from PrestaShop and formatting it to cents before sending it to Adyen servers.