Add Redirect endpoint to forward the POST parameters to the Result url
Because of the session cookie's same site attribute settings the module
cannot receive POST requests from external domains using the session
cookie. To avoid using the session cookie the Redirect controller cannot
extend the FrontController or Controller (abstract) classes since those
would break when loading the current cart based on the session cookie.

This endpoints only purpose is to receive the MD and PaRes POST
parameters from the issuer and POST it towards the Result endpoint of
the Adyen module.