Fix validation for payment methods
This fixes an issue in the Checkout page for the plugins. When one of the payment methods is set to valid, you could click the payment button or select a different payment method and finish the transaction without triggering the validation a second time.

This commit changes the flag that allows a payment to proceed to an object that holds states for different payment methods.